+++ com.unity.visualeffectgraph/Editor/Models/Operators/Implementations/Condition.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Collections.Generic;
 using System.Linq;
 using UnityEngine;
 using UnityEngine.VFX;
@@ -6,18 +7,10 @@ using UnityEngine.VFX;
 namespace UnityEditor.VFX.Operator
 {
     [VFXInfo(category = "Logic")]
-    class Condition : VFXOperator
+    class Condition : VFXOperatorDynamicType
     {
         [VFXSetting, SerializeField, Tooltip("Specifies the comparison condition between the Left and Right operands.")]
         protected VFXCondition condition = VFXCondition.Equal;
-
-        public class InputProperties
-        {
-            [Tooltip("Sets the left operand which will be compared to the right operand based on the specified condition.")]
-            public float left = 0.0f;
-            [Tooltip("Sets the right operand which will be compared to the left operand based on the specified condition.")]
-            public float right = 0.0f;
-        }
 
         public class OutputProperties
         {
@@ -27,9 +20,33 @@ namespace UnityEditor.VFX.Operator
 
         override public string name { get { return "Compare"; } }
 
+        protected override IEnumerable<VFXPropertyWithValue> inputProperties
+        {
+            get
+            {
+                yield return new VFXPropertyWithValue(
+                    new VFXProperty(GetOperandType(), "left", new TooltipAttribute("Sets the left operand which will be compared to the right operand based on the specified condition.")),
+                    GetDefaultValueForType(GetOperandType()));
+                yield return new VFXPropertyWithValue(
+                    new VFXProperty(GetOperandType(), "right", new TooltipAttribute("Sets the right operand which will be compared to the left operand based on the specified condition.")),
+                    GetDefaultValueForType(GetOperandType()));
+            }
+        }
+
+        public override IEnumerable<int> staticSlotIndex => Enumerable.Empty<int>();
+
+        public override IEnumerable<Type> validTypes => new[]
+        {
+            typeof(float),
+            typeof(uint),
+            typeof(int),
+        };
+
+        protected override Type defaultValueType => typeof(float);
+
         protected override sealed VFXExpression[] BuildExpression(VFXExpression[] inputExpression)
         {
-            return new[] { new VFXExpressionCondition(condition, inputExpression[0], inputExpression[1]) };
+            return new[] { new VFXExpressionCondition(VFXExpression.GetVFXValueTypeFromType(GetOperandType()), condition, inputExpression[0], inputExpression[1]) };
         }
     }
 }